+++ test/test_helper.rb
@@ -5,6 +5,8 @@ ActiveRecord::Migrator.migrations_paths = [File.expand_path("../test/dummy/db/mi
 require "rails/test_help"
 require "capybara/rails"
 
+Vident::StableId.strategy = Vident::StableId::RANDOM_FALLBACK
+
 # Load fixtures from the engine
 if ActiveSupport::TestCase.respond_to?(:fixture_paths=)
   ActiveSupport::TestCase.fixture_paths = [File.expand_path("fixtures", __dir__)]